feat: Add script for modifying global_prep_cmd setting in sunshine.conf file
This commit adds a PowerShell script that modifies the global_prep_cmd setting in the sunshine.conf configuration file to add a command that runs MonitorSwap-Dummy.ps1.

The script contains functions for getting the current value of global_prep_cmd from the configuration file, removing any existing commands that contain MonitorSwap-Dummy from the global_prep_cmd value, setting a new value for global_prep_cmd in the configuration file, and adding a new command to run MonitorSwap-Dummy.ps1 to the global_prep_cmd value.

The script also includes logic to check if the current user has administrator privileges, and if not, re-launches the script with elevated privileges